Changed sound data types to float and int, bigger precision is not needed Refactored, moved lfo work into oscillator and gain

+++ trunk/java/src/andts/javasynth/generator/SamplePreAmplifier.java
@@ -2,7 +2,7 @@ package andts.javasynth.generator;
 
 /**
  * Pre-amplifies signal from oscillator:
- * translates double value from oscillator into a real sample value in long
+ * translates float value from oscillator into a real sample value in int
  * amp factor depends on sample resolution
  */
 public class SamplePreAmplifier
@@ -14,7 +14,7 @@ public class SamplePreAmplifier
         ampFactor = (int) Math.round(Math.pow(2, sampleSize - 1));
     }
 
-    public long getAmplifiedValue(double value)
+    public int getAmplifiedValue(float value)
     {
         return Math.round(value * ampFactor);
     }
